rss/requirements.txt
boto3
feedparser
python-telegram-bot
markdownify